backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
GitPython>=3.1.43
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
beautifulsoup4>=4.12.3
requests>=2.32.3
SQLAlchemy>=2.0.31
pydantic-settings>=2.3.4
